\name{schneider}
\alias{schneider}
\non_function{}

\title{Quasispecies variation}

\description{
  Quasispecies variation in CMV and TMV populations in different host
  species.
}

\usage{data(schneider)}

\format{
  A data frame with six columns: virus type (CMV/TMV), host species,
  number of mutated clones, total number of clones, number of mutations,
  and total number of bases.
}

%\details{
%}

\references{
Schneider, W. L. and Roossinck, M. J. (2001) Genetic diversity in RNA
virus quasispecies is controlled by host-virus interactions.
\emph{J. Virol.} \bold{75}, 6566-6571. (See Table 1.)
}

%\source{
%}

\examples{
data(schneider)
}

\keyword{datasets}
